services/document/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
pytesseract==0.3.10
